refactor: DSU and Partition
- change constructors from int to size_t
- simplify tests accordingly
- remove get_dsu from Partition not used
- better naming for find function in Partition
- remove redundant for loop after identifying roots

+++ include/mags/DisjointSetUnion.h
@@ -11,7 +11,7 @@ class DisjointSetUnion {
 
 public:
   DisjointSetUnion() = default;
-  explicit DisjointSetUnion(int n);
+  explicit DisjointSetUnion(size_t n);
 
   [[nodiscard]] NodeID find(NodeID node) const;
